enemy attacks and patrols
Added an inheritance system. Enemies now auto-patrol within an area when the game starts. They attack if the player becomes agressiveand follow the player. If the  player gets too close when agressive the enemy attacks. I will be reworing the health system for the enemies this weekend and finishing the audio hopefully.